Create remove duplicates from a sorted array.java
INTUITION

My first thought was to iterate through the array and compare elements. If the compared elements are different, we increment the unique element counter. if not then we copy the element after the next element that we used to compare with our current iteration. But what happens when there are more than 2 of the same element. Also what happens when we hit the last index? we can't compare with something that isn't there.
I struggled a bit with what to do if there were more than 2 of the same element. So I figured if I could store the index of the last unique element into a variable and just go over each repeating element and increment the number of "skipped indexes", then I could keep track of the skipped elements and insert the next unique element into the index after the last unique element.

APPROACH

I needed three variables. One to count how many unique elements there are, one to count how many indexes I skipped. and one to store the difference between the first repeated element and the number of elements skipped. But first, we check if the array is empty. If so we return 0. If it isn't empty, we iterate through the array with a for loop. Then at the ith iteration where the next occurence of a unique element occurs we copy that value into the index after the last unique element.